Moved grpc client
